Added PUT api

+++ src/main/java/com/example/demo/categories/CategoriesService.java
@@ -2,8 +2,11 @@ package com.example.demo.categories;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -46,16 +49,33 @@ public class CategoriesService {
     }
   }
 
-  public void updateCategory(Categories categories, Long id) {
-    for (Categories i : categoriesList) {
-      if (i.getId() == id) {
-        i.setname("New name");
+  public ResponseEntity<Categories> updateCategory(Categories categories, Long id) {
+      // Check if the resource with the given ID exists
+      Optional<Categories> existingProduct = categoriesRepository.findById(id);
+      if (existingProduct.isPresent()) {
+        Categories c = existingProduct.get();
+        // Update the product with the new data
+        c.setname(categories.getname());
+        // Save the updated product to the database
+        Categories savedCategory = categoriesRepository.save(c);
+
+        return ResponseEntity.ok(savedCategory); // Return the updated product with a 200 OK response
+      } else {
+        return ResponseEntity.notFound().build(); // 404 Not Found if the resource doesn't exist
       }
-    }
+    
   }
 
-  public void deleteCategory(Categories categories, Long id) {
+  public boolean deleteCategory(Long id) {
+    // Check if the category exists
+    if (!categoriesRepository.existsById(id)) {
+      return false;
+    }
+    // Perform the DELETE operation
     categoriesRepository.deleteById(id);
+
+    // Return a 204 No Content response
+    return true;
   }
 
 }
